baclog/requirements.txt
fabric
PyYAML